Move AuthoritySignInfo to sui-types-verified
AuthoritySignInfo now lives in sui-types-verified so Verus can attach spec
functions (auth_sig_epoch_spec, auth_sig_authority_spec) without orphan
issues. sui-types re-exports the type so all existing imports are unchanged.
