+++ crates/sui-types/verified/Cargo.toml
@@ -13,10 +13,13 @@ verify = true
 
 [dependencies]
 anyhow.workspace = true
+bcs.workspace = true
 derive_more = { workspace = true, features = ["as_ref"] }
 fastcrypto.workspace = true
 schemars.workspace = true
 serde.workspace = true
+serde-name.workspace = true
 serde_with.workspace = true
+shared-crypto.workspace = true
 sui-sdk-types.workspace = true
 vstd.workspace = true
